Refactored song components and related routing
