--- src/components/songs-page/songs-page.tsx
@@ -1,38 +1,0 @@
-import { useEffect, useState } from 'react'
-import { Route, Routes } from 'react-router-dom'
-
-// FPCC
-import { FVSong } from 'components/common/data/types'
-import fetchSongsData from 'services/songsApiService'
-import { LoadingSpinner } from 'components/common/loading-spinner/loading-spinner'
-import SongsView from 'components/songs-view/songs-view'
-import SongPage from 'components/song-page/song-page'
-
-export default function SongsPage() {
-  const [songsData, setSongsData] = useState<FVSong[]>([])
-  const [loading, setLoading] = useState(true)
-
-  useEffect(() => {
-    const fetchDataAsync = async () => {
-      try {
-        const result = await fetchSongsData()
-        setSongsData(result)
-        setLoading(false)
-      } catch (error) {
-        // Handle error scenarios
-        console.error(error)
-      }
-    }
-
-    fetchDataAsync()
-  }, [])
-
-  return loading ? (
-    <LoadingSpinner />
-  ) : (
-    <Routes>
-      <Route path="" element={<SongsView songsData={songsData} />} />
-      <Route path=":id" element={<SongPage songsData={songsData} />} />
-    </Routes>
-  )
-}
